Added validation and comments

+++ routes/event.js
@@ -6,8 +6,18 @@ const Event = require("../models/event");
 
 const router = express.Router();
 
-router.get("/", (req, res, next) => {
-  Event.findOne()
+/* Get Single Event Endpoint  */
+router.get("/:id", (req, res, next) => {
+  const id = req.params.id;
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
+
+  Event.findOne(id)
     .sort({ createdAt: "desc" })
     .then(events => {
       res.json(events);
@@ -17,7 +27,12 @@ router.get("/", (req, res, next) => {
     });
 });
 
+/* Get All Events Endpoint  */
+
 router.get("/all", (req, res, next) => {
+  /* Validation */
+
+  /*            */
   Event.find()
     .sort({ createdAt: "desc" })
     .then(events => {
@@ -28,9 +43,39 @@ router.get("/all", (req, res, next) => {
     });
 });
 
-router.post("/", (req, res, next) => {
-  const newEvent = req.body;
+/* Post New Event Endpoint  */
 
+router.post("/", (req, res, next) => {
+  const { name, description, location, date, contact } = req.body;
+  /* Validation */
+  if (!name) {
+    const err = new Error("The `name` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (!description) {
+    const err = new Error("The `description` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (!location) {
+    const err = new Error("The `location` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (!date) {
+    const err = new Error("The `date` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (!contact) {
+    const err = new Error("The `contact` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
+
+  const newEvent = { name, description, location, date, contact };
   Event.create(newEvent)
     .then(response => {
       res.json(response);
@@ -40,12 +85,65 @@ router.post("/", (req, res, next) => {
     });
 });
 
+/* Put/Edit Event Endpoint  */
+
 router.put("/", (req, res, next) => {
   const { followId, name, location, description, contact, date } = req.body;
+  let event = {};
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(followId)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (name) {
+    if (typeof name !== String) {
+      const err = new Error("The `name` is not valid");
+      err.status = 400;
+      return next(err);
+    } else {
+      event.name = name;
+    }
+  }
+  if (description) {
+    if (typeof description !== String) {
+      const err = new Error("The `name` is not valid");
+      err.status = 400;
+      return next(err);
+    } else {
+      event.description = description;
+    }
+  }
+  if (location) {
+    if (typeof location !== String) {
+      const err = new Error("The `name` is not valid");
+      err.status = 400;
+      return next(err);
+    } else {
+      event.location = location;
+    }
+  }
+  if (date) {
+    if (typeof date !== String) {
+      const err = new Error("The `name` is not valid");
+      err.status = 400;
+      return next(err);
+    } else {
+      event.date = date;
+    }
+  }
+  if (contact) {
+    if (typeof contact !== String) {
+      const err = new Error("The `name` is not valid");
+      err.status = 400;
+      return next(err);
+    } else {
+      event.contact = contact;
+    }
+  }
+  /*            */
 
-  const updated = { name, location, description, contact, date };
-
-  Event.findOneAndUpdate({ _id: followId }, { updated })
+  Event.findOneAndUpdate({ _id: followId }, { event })
     .sort({ createdAt: "desc" })
     .then(event => {
       res.json(event);
@@ -55,8 +153,17 @@ router.put("/", (req, res, next) => {
     });
 });
 
+/* Delete Single Event Endpoint  */
+
 router.delete("/", (req, res, next) => {
   const id = req.body;
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
   Event.findOneAndDelete({ _id: id })
     .then(event => {
       res.json(event);
