+++ routes/following.js
@@ -6,7 +6,33 @@ const Following = require("../models/following");
 
 const router = express.Router();
 
-router.get("/", (req, res, next) => {
+/* Get Single Follow Endpoint  */
+
+router.get("/:id", (req, res, next) => {
+  const id = req.params.id;
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
+  Following.findOne(id)
+    .sort({ createdAt: "desc" })
+    .then(follows => {
+      res.json(follows);
+    })
+    .catch(err => {
+      next(err);
+    });
+});
+
+/* Get All Follows Endpoint  */
+
+router.get("/all", (req, res, next) => {
+  /* Validation */
+
+  /*            */
   Following.findOne()
     .sort({ createdAt: "desc" })
     .then(follows => {
@@ -17,9 +43,23 @@ router.get("/", (req, res, next) => {
     });
 });
 
-router.post("/", (req, res, next) => {
-  const newFollow = req.body;
+/* Post New Follow Endpoint  */
 
+router.post("/", (req, res, next) => {
+  const { userId, orgId } = req.body;
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(userId)) {
+    const err = new Error("The `User id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (!mongoose.Types.ObjectId.isValid(orgId)) {
+    const err = new Error("The `Organization id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
+  const newFollow = { userId, following: true, orgId };
   Following.create(newFollow)
     .then(response => {
       res.json(response);
@@ -29,9 +69,22 @@ router.post("/", (req, res, next) => {
     });
 });
 
+/* Put/Edit Follow Endpoint  */
+
 router.put("/", (req, res, next) => {
   const { id, following } = req.body;
-
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (typeof following !== Boolean) {
+    const err = new Error("The `Follow input` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
   Following.findOneAndUpdate({ _id: followId }, { following })
     .sort({ createdAt: "desc" })
     .then(follow => {
@@ -42,8 +95,17 @@ router.put("/", (req, res, next) => {
     });
 });
 
+/* Delete Single Follow Endpoint  */
+
 router.delete("/", (req, res, next) => {
   const id = req.body;
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
   Following.findOneAndDelete({ _id: id })
     .then(follow => {
       res.json(follow);
