+++ routes/rsvp.js
@@ -6,8 +6,18 @@ const Rsvp = require("../models/rsvp");
 
 const router = express.Router();
 
-router.get("/", (req, res, next) => {
-  Rsvp.findOne()
+/* Get Single Rsvp Endpoint  */
+
+router.get("/:id", (req, res, next) => {
+  const id = req.params.id;
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
+  Rsvp.findOne(id)
     .sort({ createdAt: "desc" })
     .then(rsvps => {
       res.json(rsvps);
@@ -17,7 +27,12 @@ router.get("/", (req, res, next) => {
     });
 });
 
+/* Get All Rsvps Endpoint  */
+
 router.get("/all", (req, res, next) => {
+  /* Validation */
+
+  /*            */
   Rsvp.find()
     .sort({ createdAt: "desc" })
     .then(rsvps => {
@@ -28,8 +43,33 @@ router.get("/all", (req, res, next) => {
     });
 });
 
+/* Post New Rsvp Endpoint  */
+
 router.post("/", (req, res, next) => {
-  const newRsvp = req.body;
+  const { userId, orgId, rsvp } = req.body;
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(userId)) {
+    const err = new Error("The `User id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (!mongoose.Types.ObjectId.isValid(orgId)) {
+    const err = new Error("The `Organization id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (!rsvp) {
+    const err = new Error("Missing `rsvp` in request body");
+    err.status = 400;
+    return next(err);
+  }
+  if (typeof rsvp !== String) {
+    const err = new Error("Error 'rsvp' was not a string");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
+  const newRsvp = { userId, rsvp, orgId };
 
   Rsvp.create(newRsvp)
     .then(response => {
@@ -40,9 +80,27 @@ router.post("/", (req, res, next) => {
     });
 });
 
+/* Put/Edit Rsvp Endpoint  */
+
 router.put("/", (req, res, next) => {
   const { rsvpId, rsvp } = req.body;
-
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(rsvpId)) {
+    const err = new Error("The `role id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  if (!rsvpId) {
+    const err = new Error("Missing `rsvpId` in request body");
+    err.status = 400;
+    return next(err);
+  }
+  if (typeof rsvpId !== String) {
+    const err = new Error("Error 'rsvpId' was not a string");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
   Rsvp.findOneAndUpdate({ _id: rsvpId }, { rsvp })
     .sort({ createdAt: "desc" })
     .then(rsvp => {
@@ -53,8 +111,17 @@ router.put("/", (req, res, next) => {
     });
 });
 
+/* Delete Single Rsvp Endpoint  */
+
 router.delete("/", (req, res, next) => {
   const id = req.body;
+  /* Validation */
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
+  /*            */
   Rsvp.findOneAndDelete({ _id: id })
     .then(rsvp => {
       res.json(rsvp);
